Redoing, mostly. Some new stuff, I guess.
