+++ src/board/mod.rs
@@ -1,6 +1,8 @@
 pub mod line;
 pub mod gate;
+pub mod basics;
+pub mod simplegate;
+pub mod mux;
+pub mod memory;
 pub mod circuit;
 //pub mod board;
-pub mod basics;
-pub mod memory;
